Move auth & dashboard UI into modules
Refactor pages to use modular components and reorganize UI code for reuse. Large page implementations (auth, collaborations, communities, discover, ai-chat, etc.) were replaced with imports of new module components (e.g. LoginForm, CollaborationManager, CommunityControlCenter, ChummeAIChatPage). Many new components, types and mock-data files were added under modules/* (auth, collaboration, community, dashboard, discover, entertainment, landing). Shared UI pieces were moved/renamed into module paths (AuthCard, AuthLayout, AuthInitializer, Navbar). Also added agent.md and info.md. Purpose: simplify page files, centralize UI logic, and prepare the codebase for easier maintenance and feature development.
