+++ app/error.tsx
@@ -3,7 +3,7 @@
 import { useEffect } from "react";
 import { Button } from "@/modules/shared/components/Button";
 import { Card } from "@/modules/shared/components/Card";
-import { Navbar } from "@/modules/shared/components/Navbar";
+import { Navbar } from "@/modules/landing/components/Navbar";
 
 export default function Error({
   error,
