+++ modules/auth/components/AuthComponents.tsx
@@ -1,0 +1,12 @@
+
+
+
+const AuthComponents = () => {
+    return (
+        <div>
+            <h1>AuthComponents</h1>
+        </div>
+    );
+};
+
+export default AuthComponents;
