+++ modules/collaboration/constants/mock-data.ts
@@ -1,0 +1,119 @@
+import type { CollabRoom, PublicCollab, Song, Recording } from "@/modules/collaboration/types";
+
+export const mockRooms: CollabRoom[] = [
+  {
+    id: "1",
+    name: "Kpop Night Room",
+    type: "Public",
+    host: "user_123",
+    participants: 8,
+    song: "Dynamite",
+    status: "Active",
+  },
+  {
+    id: "2",
+    name: "Fan Singing PH",
+    type: "Private",
+    host: "maria23",
+    participants: 5,
+    song: "How You Like That",
+    status: "Active",
+  },
+  {
+    id: "3",
+    name: "BTS Sing Along",
+    type: "Public",
+    host: "armyfan",
+    participants: 12,
+    song: "Butter",
+    status: "Active",
+  },
+  {
+    id: "4",
+    name: "Private Jam Session",
+    type: "Private",
+    host: "john_doe",
+    participants: 3,
+    song: "Kill This Love",
+    status: "Ended",
+  },
+];
+
+export const mockPublicCollabs: PublicCollab[] = [
+  {
+    id: "1",
+    title: "Global K-Pop Collab",
+    creator: "sarah_kim",
+    category: "K-Pop",
+    participants: 45,
+    startDate: "Mar 8, 2024",
+    status: "Approved",
+  },
+  {
+    id: "2",
+    title: "Taylor Swift Night",
+    creator: "swiftie_fan",
+    category: "Pop",
+    participants: 32,
+    startDate: "Mar 9, 2024",
+    status: "Pending",
+  },
+  {
+    id: "3",
+    title: "Anime OST Covers",
+    creator: "anime_lover",
+    category: "Anime",
+    participants: 28,
+    startDate: "Mar 7, 2024",
+    status: "Approved",
+  },
+];
+
+export const mockSongs: Song[] = [
+  { id: "1", title: "Dynamite", artist: "BTS", usageCount: 230, status: "Active" },
+  {
+    id: "2",
+    title: "How You Like That",
+    artist: "BLACKPINK",
+    usageCount: 180,
+    status: "Active",
+  },
+  { id: "3", title: "Butter", artist: "BTS", usageCount: 195, status: "Active" },
+  {
+    id: "4",
+    title: "Kill This Love",
+    artist: "BLACKPINK",
+    usageCount: 165,
+    status: "Active",
+  },
+];
+
+export const mockRecordings: Recording[] = [
+  {
+    id: "1",
+    title: "Amazing K-Pop Cover",
+    room: "Kpop Night Room",
+    creator: "user_123",
+    duration: "3:45",
+    uploadDate: "Mar 8, 2024",
+    featured: true,
+  },
+  {
+    id: "2",
+    title: "BTS Medley",
+    room: "BTS Sing Along",
+    creator: "armyfan",
+    duration: "5:20",
+    uploadDate: "Mar 7, 2024",
+    featured: false,
+  },
+  {
+    id: "3",
+    title: "Girls Group Mashup",
+    room: "Fan Singing PH",
+    creator: "maria23",
+    duration: "4:10",
+    uploadDate: "Mar 9, 2024",
+    featured: false,
+  },
+];
