+++ modules/discover/components/LinksTab.tsx
@@ -1,0 +1,163 @@
+"use client";
+
+import { discoverLinks } from "@/modules/discover/constants/mock-data";
+
+interface LinksTabProps {
+  isDarkMode: boolean;
+}
+
+export function LinksTab({ isDarkMode }: LinksTabProps) {
+  return (
+    <div>
+      <div className="flex justify-between items-center mb-6">
+        <h2
+          className={`text-xl font-bold ${isDarkMode ? "text-white" : "text-gray-900"}`}
+        >
+          Discover Links
+        </h2>
+        <button className="px-4 py-2 rounded-lg bg-gradient-to-r from-[#A53860] to-[#670D2F] text-white font-medium hover:opacity-90 transition-opacity">
+          Add Link
+        </button>
+      </div>
+
+      <div
+        className={`rounded-lg border overflow-hidden ${
+          isDarkMode ? "bg-gray-800 border-gray-700" : "bg-white border-gray-200"
+        }`}
+      >
+        <div className="overflow-x-auto">
+          <table className="w-full">
+            <thead className={isDarkMode ? "bg-gray-900" : "bg-gray-50"}>
+              <tr>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Link Preview
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  User
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Fandom
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Platform
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Status
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Date
+                </th>
+                <th
+                  className={`px-6 py-3 text-left text-xs font-medium uppercase tracking-wider ${
+                    isDarkMode ? "text-gray-400" : "text-gray-500"
+                  }`}
+                >
+                  Actions
+                </th>
+              </tr>
+            </thead>
+            <tbody
+              className={`divide-y ${isDarkMode ? "divide-gray-700" : "divide-gray-200"}`}
+            >
+              {discoverLinks.map((link) => (
+                <tr
+                  key={link.id}
+                  className={isDarkMode ? "hover:bg-gray-700" : "hover:bg-gray-50"}
+                >
+                  <td className={`px-6 py-4 ${isDarkMode ? "text-white" : "text-gray-900"}`}>
+                    {link.preview}
+                  </td>
+                  <td
+                    className={`px-6 py-4 whitespace-nowrap ${
+                      isDarkMode ? "text-gray-300" : "text-gray-700"
+                    }`}
+                  >
+                    @{link.user}
+                  </td>
+                  <td
+                    className={`px-6 py-4 whitespace-nowrap ${
+                      isDarkMode ? "text-gray-300" : "text-gray-700"
+                    }`}
+                  >
+                    {link.fandom}
+                  </td>
+                  <td
+                    className={`px-6 py-4 whitespace-nowrap ${
+                      isDarkMode ? "text-gray-300" : "text-gray-700"
+                    }`}
+                  >
+                    {link.platform}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <span
+                      className={`px-2 py-1 text-xs rounded-full ${
+                        link.status === "Approved"
+                          ? "bg-green-100 text-green-800"
+                          : "bg-yellow-100 text-yellow-800"
+                      }`}
+                    >
+                      {link.status}
+                    </span>
+                  </td>
+                  <td
+                    className={`px-6 py-4 whitespace-nowrap ${
+                      isDarkMode ? "text-gray-300" : "text-gray-700"
+                    }`}
+                  >
+                    {link.date}
+                  </td>
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="flex items-center gap-2">
+                      <button
+                        className={`px-3 py-1 text-sm rounded ${
+                          isDarkMode
+                            ? "bg-gray-700 text-white hover:bg-gray-600"
+                            : "bg-gray-100 text-gray-700 hover:bg-gray-200"
+                        }`}
+                      >
+                        Edit
+                      </button>
+                      <button
+                        className={`px-3 py-1 text-sm rounded ${
+                          isDarkMode
+                            ? "bg-red-900 text-red-200 hover:bg-red-800"
+                            : "bg-red-100 text-red-700 hover:bg-red-200"
+                        }`}
+                      >
+                        Delete
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      </div>
+    </div>
+  );
+}
